backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg[binary]>=3.1
requests>=2.31
arabic-reshaper>=3.0
python-bidi>=0.4
pydub>=0.25
celery>=5.3
redis>=5.0
gunicorn>=21.2
google-cloud-texttospeech>=2.21.0
